Rework nestegg_track_codec_data to use io_buffer.
This was encouraged by a scan-build report for errant behaviour in this
code that I failed to debug but which disappears with this
rewrite (hopefully because it's fixed, rather than because the analyzer
is no longer able to see deep enough to recognize a bug).  As a bonus,
this avoids duplicating Xiph unlacing in two places and a bunch of raw
pointer arithmetic.