Improve test compatibility for MinGW builds.
Looks like there was a typo in endian detection which would cause mingw to try
and import endian.h, which would fail. This typo has been fixed so now the
appropriate preprocessor define will be checked.

Updated the way comparisons are done by the tests to use diff instead of cmp,
which lets line endings be stripped. Previously, line endings differences would
result in failures, even when the rest of the test output matched, this
prevents such failures.