Avoid clang warning about uninitialized data_size
When building manually for iOS with Xcode, clang complains
about a possibly uninitialized data_size at the end of the
loop in ne_read_block_additions.

This is already safe because it's always initialized when
has_data is set to 1, and never used when has_data is left
at 0, but initializing data_size to 0 along with has_data
at the head of the loop silences the warning.

Should have no impact on functionality but keeps the compiler
happier.