Fix EOF handling in ne_buffer_read (used by nestegg_sniff).
Previously, any short read would be treated as EOF.  Now, EOF is only
returned if the entire buffer has been consumed, and any short read is
signalled as an error.  This matches the expected behaviour of the
nestegg_io interface.

This code is duplicated in test/fuzz.cc, so also fixed there.

This addresses an OOM found by the oss-fuzz project (issue 8875).