Add new public API `nestegg_read_last_packet()`
In gecko, AacCodecSpecificData has an attribute mMediaFrameCount, which
is calculated by

> Media frame count = total frame count - delay frame count - padding frame count.

Nestegg already has methods to get the delay frame amount (codec_delay
in nestegg_audio_params) and the padding frame amount (
nestegg_packet_discard_padding).

In order to know the total frame count, we have to know the timestamp of
the last packet in AAC stream. Therefore, we add a new method
nestegg_read_last_packet to achieve that.